Add go dep and run goimports

+++ api/v1/users.go
@@ -1,13 +1,13 @@
 package v1
 
 import (
-	"../../miniserver"
 	"net/http"
-
-	"../../database"
-	"../../utils"
 	"strconv"
-	"../../logger"
+
+	"github.com/Grarak/GoYTFetcher/database"
+	"github.com/Grarak/GoYTFetcher/logger"
+	"github.com/Grarak/GoYTFetcher/miniserver"
+	"github.com/Grarak/GoYTFetcher/utils"
 )
 
 func usersSignUp(client *miniserver.Client) miniserver.Response {
@@ -49,8 +49,7 @@ func usersList(client *miniserver.Client) miniserver.Response {
 	}
 
 	usersDB := database.GetDatabase().UsersDB
-	if user, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *user.Verified {
+	if user, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *user.Verified {
 		page, err := strconv.Atoi(client.Queries.Get("page"))
 		if err != nil {
 			page = 1
@@ -71,8 +70,7 @@ func usersSetVerification(client *miniserver.Client) miniserver.Response {
 	}
 
 	usersDB := database.GetDatabase().UsersDB
-	if requester, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Admin {
+	if requester, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Admin {
 		err = usersDB.SetVerificationUser(request)
 		if err == nil {
 			return client.CreateResponse(utils.StatusNoError)
@@ -89,8 +87,7 @@ func usersDelete(client *miniserver.Client) miniserver.Response {
 	}
 
 	usersDB := database.GetDatabase().UsersDB
-	if requester, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Admin {
+	if requester, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Admin {
 		err = usersDB.DeleteUser(request)
 		if err != nil {
 			return client.CreateResponse(utils.StatusNoError)
@@ -107,8 +104,7 @@ func usersDeleteAll(client *miniserver.Client) miniserver.Response {
 	}
 
 	usersDB := database.GetDatabase().UsersDB
-	if requester, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Admin {
+	if requester, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Admin {
 		err = usersDB.DeleteAllNonVerifiedUsers(request)
 		if err == nil {
 			return client.CreateResponse(utils.StatusNoError)
@@ -125,8 +121,7 @@ func usersResetPassword(client *miniserver.Client) miniserver.Response {
 	}
 
 	usersDB := database.GetDatabase().UsersDB
-	if requester, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Admin {
+	if requester, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Admin {
 		err = usersDB.ResetPasswordUser(request)
 		if err == nil {
 			return client.CreateResponse(utils.StatusNoError)
@@ -159,8 +154,7 @@ func playlistListPublic(client *miniserver.Client) miniserver.Response {
 
 	usersDB := database.GetDatabase().UsersDB
 	playlistsDB := database.GetDatabase().PlaylistsDB
-	if requester, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Verified {
+	if requester, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Verified {
 
 		user, err := usersDB.FindUserByName(request.Name)
 		if err == nil {
@@ -182,8 +176,7 @@ func playlistCreate(client *miniserver.Client) miniserver.Response {
 
 	usersDB := database.GetDatabase().UsersDB
 	playlistsDB := database.GetDatabase().PlaylistsDB
-	if requester, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Verified {
+	if requester, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Verified {
 
 		err := playlistsDB.CreatePlaylist(request)
 		if err == nil {
@@ -247,8 +240,7 @@ func playlistListIdsPublic(client *miniserver.Client) miniserver.Response {
 
 	usersDB := database.GetDatabase().UsersDB
 	playlistsDB := database.GetDatabase().PlaylistsDB
-	if requester, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Verified {
+	if requester, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Verified {
 		user, err := usersDB.FindUserByName(request.Name)
 		if err == nil {
 			playlist := database.Playlist{ApiKey: user.ApiKey, Name: request.Playlist}
@@ -272,8 +264,7 @@ func playlistAddId(client *miniserver.Client) miniserver.Response {
 
 	usersDB := database.GetDatabase().UsersDB
 	playlistsDB := database.GetDatabase().PlaylistsDB
-	if requester, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Verified {
+	if requester, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Verified {
 		err = playlistsDB.AddIdToPlaylist(request)
 		if err != nil {
 			return client.CreateResponse(utils.StatusPlaylistIdAlreadyExists)
@@ -295,8 +286,7 @@ func playlistDeleteId(client *miniserver.Client) miniserver.Response {
 
 	usersDB := database.GetDatabase().UsersDB
 	playlistsDB := database.GetDatabase().PlaylistsDB
-	if requester, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Verified {
+	if requester, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Verified {
 		err := playlistsDB.DeleteIdFromPlaylist(request)
 		if err == nil {
 			return client.CreateResponse(utils.StatusNoError)
@@ -314,8 +304,7 @@ func playlistSetIds(client *miniserver.Client) miniserver.Response {
 
 	usersDB := database.GetDatabase().UsersDB
 	playlistsDB := database.GetDatabase().PlaylistsDB
-	if requester, err := usersDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Verified {
+	if requester, err := usersDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Verified {
 		err := playlistsDB.SetPlaylistIds(request)
 		if err == nil {
 			return client.CreateResponse(utils.StatusNoError)
@@ -333,8 +322,7 @@ func historyAdd(client *miniserver.Client) miniserver.Response {
 
 	userDB := database.GetDatabase().UsersDB
 	historiesDB := database.GetDatabase().HistoriesDB
-	if requester, err := userDB.FindUserByApiKey(request.ApiKey);
-		err == nil && *requester.Verified {
+	if requester, err := userDB.FindUserByApiKey(request.ApiKey); err == nil && *requester.Verified {
 		err = historiesDB.AddHistory(request.ApiKey, request.Id)
 		if err == nil {
 			logger.I(client.IPAddr + ": " + requester.Name +
